requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.82
botocore==1.34.82
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.14
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-debug-toolbar-request-history==0.1.4
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.15.1
frozenlist==1.4.1
geoip2==4.8.0
idna==3.6
jmespath==1.0.1
maxminddb==2.6.2
multidict==6.0.5
pillow==10.3.0
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
stripe==9.2.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
